+++ src/elements/actions/ImportEpisodeFromRSS.php
@@ -50,7 +50,8 @@ class ImportEpisodeFromRSS extends ElementAction
             return true;
         },
         activate: \$selectedItems => {
-            window.open('episodes/import-from-rss?podcastId=' + \$selectedItems.find('.element').data('id'), '_self');
+            window.open('episodes/import-from-rss?podcastId=' + \$selectedItems.find('.element').data('id') + '&siteId=' +
+            \$selectedItems.find('.element').data('siteId'), '_self');
         },
     });
 })();