+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added importerService
 - Added support for propagate-to and set-enabled-for-site for console resave
 - Added support for different podcast general settings per site
+- Added importing episodes to specific sites from asset indexes and RSS
 - Now each podcast has two settings pages (general settings, episode settings)
 - Episode summary and description can be different now
 - Added basic support for pubDate on podcast RSS